+++ ble_uart.h
@@ -1,0 +1,9 @@
+#ifndef GHOST_BLE_UART_H
+#define GHOST_BLE_UART_H
+
+#include <bluefruit.h>
+
+void setupBleUart();
+void handleBleUart();
+
+#endif // GHOST_BLE_UART_H